system/file.go: add IsSamePath, IsFilePath and IsDirPath, add TestIsSamePath.

+++ internal/system/file.go
@@ -1,6 +1,7 @@
 package system
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -37,6 +38,13 @@ func WriteFile(filename string, data []byte) error {
 
 // CopyFile is used to copy file from source path to destination path.
 func CopyFile(dst, src string) error {
+	same, err := IsSamePath(dst, src)
+	if err != nil {
+		return err
+	}
+	if same {
+		return nil
+	}
 	srcFile, err := os.Open(src) // #nosec
 	if err != nil {
 		return err
@@ -71,14 +79,61 @@ func CopyFile(dst, src string) error {
 // MoveFile is used to move file from source path to destination path.
 // It can move file to the different volume(not use os.Rename).
 func MoveFile(dst, src string) error {
-	err := CopyFile(dst, src)
+	same, err := IsSamePath(dst, src)
+	if err != nil {
+		return err
+	}
+	if same {
+		return nil
+	}
+	err = CopyFile(dst, src)
 	if err != nil {
 		return err
 	}
 	return os.Remove(src)
 }
 
-// IsPathExist is used to check the target file or directory is exist.
+// IsSamePath is used to check paths are same(absolute).
+func IsSamePath(path ...string) (bool, error) {
+	l := len(path)
+	if l < 2 {
+		return false, errors.New("must input more than one path")
+	}
+	abs, err := filepath.Abs(path[0])
+	if err != nil {
+		return false, err
+	}
+	for i := 1; i < l; i++ {
+		a, err := filepath.Abs(path[i])
+		if err != nil {
+			return false, err
+		}
+		if a != abs {
+			return false, nil
+		}
+	}
+	return true, nil
+}
+
+// IsFilePath is used to check the target path is a file.
+func IsFilePath(path string) (bool, error) {
+	fi, err := os.Stat(path)
+	if err != nil {
+		return false, err
+	}
+	return !fi.IsDir(), nil
+}
+
+// IsDirPath is used to check the target path is a directory.
+func IsDirPath(path string) (bool, error) {
+	fi, err := os.Stat(path)
+	if err != nil {
+		return false, err
+	}
+	return fi.IsDir(), nil
+}
+
+// IsPathExist is used to check the target path is exist.
 func IsPathExist(path string) (bool, error) {
 	_, err := os.Stat(path)
 	if err == nil {
@@ -90,7 +145,7 @@ func IsPathExist(path string) (bool, error) {
 	return false, err
 }
 
-// IsPathNotExist is used to check the target file or directory is not exist.
+// IsPathNotExist is used to check the target path is not exist.
 func IsPathNotExist(path string) (bool, error) {
 	_, err := os.Stat(path)
 	if err == nil {
